requirements.txt
opencv-python==4.1.2.30
Pillow==6.2.2
numpy==1.16.6
boto3==1.12.19
